Initial commit
The current state of this SOP implementation is slightly more than
basic. It supports all commands (with the exception of
detach-inband-signature-and-message) and most flags. We will continue
development upon requests of new features - as long as they are
consistent with the SOP specification - and synchronize with
subcommands and flags used in the OpenPGP interoperability test suite.

As of the time of writing, the SOP test suite only uses `version,
generate-key, extract-cert, encrypt, decrypt, sign, verify, armor,
dearmor`, each without any flags.